refact : BlogPostList 보여주는 로직 변경
